Refactoring: separate kernels - fast version for depth 1, general version for deeper level
